requirements.txt
certifi==2019.3.9
Django==2.2.1
djangorestframework==3.9.4
mysqlclient==1.4.2.post1
pytz==2019.1
sqlparse==0.3.0
Pillow==6.0.0
gunicorn
django-rest-swagger
numpy